Move `test/scalability` to `test/performance/scheduler`.
