+++ test/performance/scheduler/runner/main.go
@@ -48,11 +48,11 @@ import (
 	kueuealpha "sigs.k8s.io/kueue/apis/kueue/v1alpha1"
 	kueue "sigs.k8s.io/kueue/apis/kueue/v1beta1"
 	utiltesting "sigs.k8s.io/kueue/pkg/util/testing"
-	"sigs.k8s.io/kueue/test/scalability/runner/controller"
-	"sigs.k8s.io/kueue/test/scalability/runner/generator"
-	"sigs.k8s.io/kueue/test/scalability/runner/recorder"
-	"sigs.k8s.io/kueue/test/scalability/runner/scraper"
-	"sigs.k8s.io/kueue/test/scalability/runner/stats"
+	"sigs.k8s.io/kueue/test/performance/scheduler/runner/controller"
+	"sigs.k8s.io/kueue/test/performance/scheduler/runner/generator"
+	"sigs.k8s.io/kueue/test/performance/scheduler/runner/recorder"
+	"sigs.k8s.io/kueue/test/performance/scheduler/runner/scraper"
+	"sigs.k8s.io/kueue/test/performance/scheduler/runner/stats"
 )
 
 var (
